Fix Duration.toLocaleString
Define the abstract operation that is expected here, and add a test that
exercises toLocaleString without caring about the contents of the
returned string.
